pyproject.toml
[project]
name = "tap-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "environs>=14.2.0",
    "google-auth>=2.40.2",
    "markdown>=3.8",
    "mysqlclient>=2.2.7",
]
